Extract simil views to partial
